agentic-rag/pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=25.12.0",
    "chromadb>=1.4.0",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-google-genai>=4.1.2",
    "langchain-tavily>=0.2.16",
    "langchain-text-splitters>=1.1.0",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
]
